- replaced js https module with axios one - connected Express framework - added "nodemon" module for server autoreloading - added server and controller - structured code
