+++ README.md
@@ -7,9 +7,6 @@ https://api.nasa.gov/
 - NASA Asteroids url:
 https://api.nasa.gov/neo/rest/v1/feed
 
-- Server entry point: 
-server.js
-
 - Start server command:
 "npm run start_server"
 
@@ -25,11 +22,26 @@ link: https://www.npmjs.com/package/dotenv
 see ".env.example" file
 
 - Moment.js library:
-link: https://momentjs.com/
+1. link: 
+https://momentjs.com/
+2. moment().day() method:
+https://www.geeksforgeeks.org/moment-js-moment-day-method/
 
 - Express framework install command:
-"npm instal express"
+"npm instal express --save"
 
-- Http library:
+- Https client:
+1. library:
 Axios
-install command: npm install axios
+2. Install command: 
+"npm install axios --save"
+
+- Automatically server reloading:
+1. Module:
+nodemon
+2. Install command
+"npm install nodemon --save"
+3. Start server command:
+"nodemon --env-file .env ./js/controller/AsteroidController.js"
+4. Manual resrarting command:
+"rs" and hit "ENTER"
