+++ package.json
@@ -4,14 +4,14 @@
   "description": "Proxy server project",
   "main": "index.js",
   "scripts": {
-    "start": "node --env-file .env ./js/index.js",
-    "start_server": "node --env-file .env ./js/server.js"
+    "start_server": "nodemon --env-file .env ./js/controller/AsteroidController.js"
   },
   "author": "Kabiakou",
   "license": "ISC",
   "dependencies": {
     "axios": "^1.6.7",
     "express": "^4.18.3",
-    "moment": "^2.30.1"
+    "moment": "^2.30.1",
+    "nodemon": "^3.1.0"
   }
 }
